casgo/ml-service/requirements.txt
# FastAPI and Web Framework Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Azure OpenAI and AI Dependencies  
openai>=1.3.0
requests>=2.31.0

# Environment and Configuration
python-dotenv>=1.0.0

# HTTP Client
httpx>=0.25.0

# Testing Dependencies (lightweight)
pytest>=7.4.0

# Development Dependencies (lightweight)
black>=23.0.0 